ContactBase, Contact6D and ContactPoint implemented, binded and working. Tests are correct and we miss some tests of the ContactPoint. Documentations included.
