+++ python/contact_6d.cpp
@@ -31,6 +31,7 @@ Eigen::Matrix<double, 6, 1> get_weights(Contact6DSettings &self) {
 }
 
 void exposeContact6D() {
+
   bp::class_<Contact6DSettings>("Contact6DSettings")
       .def_readwrite("frame_name", &Contact6DSettings::frame_name)
       .def_readwrite("gu", &Contact6DSettings::gu)
@@ -42,7 +43,7 @@ void exposeContact6D() {
       .def("__eq__", &Contact6DSettings::operator==)
       .def("__ne__", &Contact6DSettings::operator!=);
 
-  bp::class_<Contact6D>("Contact6D", bp::init<>())
+  bp::class_<Contact6D, bp::bases<ContactBase>, boost::noncopyable>("Contact6D", bp::init<>())
       .def("initialize", &Contact6D::initialize, bp::args("self", "settings"))
       .def("get_settings", &get_settings, bp::args("self"))
       .def("set_mu", &Contact6D::setMu, bp::args("self", "mu"))
@@ -55,41 +56,9 @@ void exposeContact6D() {
            bp::args("self", "half_width"))
       .def("set_surface_half_length", &Contact6D::setSurfaceHalfLength,
            bp::args("self", "half_length"))
-      .def("update_NE_matrix", &Contact6D::updateNewtonEuler,
+      .def("updateNewtonEuler", &Contact6D::updateNewtonEuler,
            bp::args("self", "CoM", "oMf"))
-      .def("uni_A", &Contact6D::uni_A,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("uni_b", &Contact6D::uni_b,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("reg_A", &Contact6D::reg_A,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("reg_b", &Contact6D::reg_b,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("fri_A", &Contact6D::fri_A,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("fri_b", &Contact6D::fri_b,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("NE_A", &Contact6D::NE_A,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("uni_rows", &Contact6D::uni_rows, bp::args("self"))
-      .def("fri_rows", &Contact6D::fri_rows, bp::args("self"))
-      .def("cols", &Contact6D::cols, bp::args("self"))
-      .def("get_frame_id", &Contact6D::getFrameID, bp::args("self"))
-      .def("toWorldForces", &Contact6D::toWorldForces, bp::args("self"))
-      .def("toCoMForces", &Contact6D::toCoMForces, bp::args("self"))
-      .def("set_frame_id", &Contact6D::setFrameID, bp::args("self"))
-      .def("applyForce", &Contact6D::applyForce, bp::args("self", "force"))
-      .def("appliedForce", &Contact6D::appliedForce,
-           bp::return_value_policy<bp::reference_existing_object>(),
-           bp::args("self"))
-      .def("get_pose", &Contact6D::getPose,
+      .def("getFrameName", &Contact6D::getFrameName,
            bp::return_value_policy<bp::reference_existing_object>(),
            bp::args("self"));
 }
