+++ python/dynacom/__init__.py
@@ -4,5 +4,8 @@ from .dynacom import (
     Contact6D,
     Contact6DSettings,
     DynaCoMSettings,
-    DynaCoM
+    DynaCoM,
+    ContactBase,
+    ContactPointSettings,
+    ContactPoint
 )  # noqa: F401
